requirements.txt
flask
Flask
PyJWT
pymongo
dnspython == 1.16.0
flask-cors ==3.0.8
